While TRUE Process Cot Request Func
Fundamental change implemented a while true loop on any resource to allow to move between the care levels, we also recalculate the likely hood of need x level after leaving y level. This allows patients to move down AND up the intensity of care required. Also allows us to implement logic to use different available cots where available. Currently implemented with .AnyOf() which I dont think applies a preferential order so if more than one available an agent could take a more complex cot than required using unecessary resource. This took a long time to get to this stage as there was an issue with properly releasing resource requests leading to massively over blocked resources. So there is a big change between this iteration and the last for this reason.