Garbage Collection and Re-roll While loop
Added very basic GC to prevent memory overloads... also added a re-roll when the weighted random choice goes out of bounds of the current list. Technically this means that the distribution percentage is out when these numbers are rolled as it happens when the calculated distribution list is longer than the available site list for the current run (when we remove sites). However the list is ordered so it is the 20th or so on site on the distribution where the error occurs and so a fractional percentage error and I am happy that a re-roll of the random dice in these marginal instances is ok. 0.9985574148874783,
 0.9989071324905139,
 0.9992787074437391,
 1.0 These are the 18th - 21st proportions of the distribution so we can see that the error accounts for around 0.001% of the population distribution.